Update omz conditions

+++ install/delta.sh
@@ -1,4 +1,2 @@
-if [ -e .gitconfig ]; then
-	mv ~/.gitconfig{,.bak}
-fi
+[ -f "$HOME/.gitconfig" ] && mv ~/.gitconfig{,.bak}
 cp ~/.local/share/${project}/configs/.gitconfig ~/.gitconfig
